Version 1.0.6

+++ src/commonMain/kotlin/ru/musintimur/instastat/common/messages/DayCount.kt
@@ -1,3 +1,3 @@
 package ru.musintimur.instastat.common.messages
 
-data class DayCount(val date: String, val count: Long)
+data class DayCount(val date: String, val count: Int)
