+++ src/jvmMain/kotlin/ru/musintimur/instastat/extensions/ExtTd.kt
@@ -4,13 +4,13 @@ import kotlinx.html.TD
 import kotlinx.html.span
 import kotlinx.html.style
 
-fun TD.setColoredCell(count: Long, diff: Long) {
+fun TD.setColoredCell(count: Int, diff: Int) {
     + "$count ("
     coloredSpan(diff)
     + ")"
 }
 
-fun TD.coloredSpan(count: Long) {
+fun TD.coloredSpan(count: Int) {
     span {
         style = when {
             count > 0 -> "color:green"
