+++ src/jvmMain/kotlin/ru/musintimur/instastat/model/entities/Comment.kt
@@ -1,8 +1,8 @@
 package ru.musintimur.instastat.model.entities
 
 data class Comment(
-    val commentId: Long,
-    val postId: Long,
+    val commentId: Int,
+    val postId: Int,
     val author: String,
     val text: String
 )
