+++ src/jvmMain/kotlin/ru/musintimur/instastat/model/entities/PeriodReportRecord.kt
@@ -2,10 +2,10 @@ package ru.musintimur.instastat.model.entities
 
 data class PeriodReportRecord(
     val title: String,
-    val count_posts: Long,
-    val count_posts_diff: Long,
-    val count_followers: Long,
-    val count_followers_diff: Long,
-    val count_followings: Long,
-    val count_followings_diff: Long
+    val count_posts: Int,
+    val count_posts_diff: Int,
+    val count_followers: Int,
+    val count_followers_diff: Int,
+    val count_followings: Int,
+    val count_followings_diff: Int
 )
