+++ src/jvmMain/kotlin/ru/musintimur/instastat/model/entities/Post.kt
@@ -3,7 +3,7 @@ package ru.musintimur.instastat.model.entities
 import java.time.LocalDateTime
 
 data class Post(
-    val postId: Long,
+    val postId: Int,
     val postUrl: String,
     val postDate: LocalDateTime?,
     val postText: String?,
