+++ src/jvmMain/kotlin/ru/musintimur/instastat/model/entities/Profile.kt
@@ -1,7 +1,7 @@
 package ru.musintimur.instastat.model.entities
 
 data class Profile(
-    val profileId: Long,
+    val profileId: Int,
     val profileName: String,
     val isActive: Boolean
 )
