+++ src/jvmMain/kotlin/ru/musintimur/instastat/repository/contracts/CommentsContract.kt
@@ -8,6 +8,6 @@ interface CommentsContract {
     suspend fun getPostComments(post: Post): List<Comment>
     suspend fun addNewComment(post: Post, commentAuthor: String, commentText: String)
     suspend fun findComment(post: Post, commentAuthor: String, commentText: String): Comment?
-    suspend fun calculateComments(post: Post): Long
+    suspend fun calculateComments(post: Post): Int
 
 }
