+++ src/jvmMain/kotlin/ru/musintimur/instastat/repository/contracts/PostsContract.kt
@@ -7,10 +7,10 @@ interface PostsContract {
 
     suspend fun getAllPosts(): List<Post>
     suspend fun getPostByUrl(url: String): Post?
-    suspend fun getPostById(postId: Long): Post?
+    suspend fun getPostById(postId: Int): Post?
     suspend fun addPost(url: String)
     suspend fun updateDateTime(url: String, datetime: LocalDateTime)
     suspend fun updateText(url: String, text: String)
-    suspend fun updateCommentsCount(url: String, count: Long)
+    suspend fun updateCommentsCount(url: String, count: Int)
 
 }
