+++ src/jvmMain/kotlin/ru/musintimur/instastat/repository/implementations/PostsRepository.kt
@@ -11,17 +11,17 @@ class PostsRepository(private val queries: InstastatDatabaseQueries): PostsContr
 
     override suspend fun getAllPosts(): List<Post> =
         queries.getAllPosts().executeAsList().map {
-            Post(it.post_id, it.post_url, it.post_date.asSqlLiteDateTime(), it.post_text, it.comments_count.toInt())
+            Post(it.post_id, it.post_url, it.post_date.asSqlLiteDateTime(), it.post_text, it.comments_count)
         }
 
     override suspend fun getPostByUrl(url: String): Post? =
         queries.getPostByUrl(url).executeAsOneOrNull()?.let {
-            Post(it.post_id, it.post_url, it.post_date.asSqlLiteDateTime(), it.post_text, it.comments_count.toInt())
+            Post(it.post_id, it.post_url, it.post_date.asSqlLiteDateTime(), it.post_text, it.comments_count)
         }
 
-    override suspend fun getPostById(postId: Long): Post? =
+    override suspend fun getPostById(postId: Int): Post? =
         queries.getPostById(postId).executeAsOneOrNull()?.let {
-            Post(it.post_id, it.post_url, it.post_date.asSqlLiteDateTime(), it.post_text, it.comments_count.toInt())
+            Post(it.post_id, it.post_url, it.post_date.asSqlLiteDateTime(), it.post_text, it.comments_count)
         }
 
     override suspend fun addPost(url: String) {
@@ -36,7 +36,7 @@ class PostsRepository(private val queries: InstastatDatabaseQueries): PostsContr
         queries.updatePostText(text, url)
     }
 
-    override suspend fun updateCommentsCount(url: String, count: Long) {
+    override suspend fun updateCommentsCount(url: String, count: Int) {
         queries.updatePostCommentsCount(count, url)
     }
 }
